requirements.txt
opencv-contrib-python==4.1.0.25
pillow==6.2.1
qrcode
